\docType{class}
\name{LapplyExecutionState-class}
\alias{LapplyExecutionState-class}
\title{Class \code{"LapplyExecutionState"}}
\usage{
  LapplyExecutionStateClass(...)
}
\description{
  An instance of this class reflects the entire lifetime of
  a dynamic parallelization.
}
\section{Extends}{
  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("LapplyExecutionState")
}
\author{
  Stefan Böhringer

  Maintainer: Stefan Böhringer
  <r-packages@s-boehringer.org>
}
\seealso{
  \code{\link{LapplyFreezer-class}},
  \code{\link{LapplyState-class}}
}
\keyword{classes}

