% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparallel_setEnable_pkg.R
\name{parallelize_setEnable}
\alias{parallelize_setEnable}
\title{Turn on/off the parallelization mechanism

This function changes the definition of
Apply/Lapply/Sapply/parallelize/parallelize_call to turn the parallelization
mechanism on or off.

This function changes the definition of
Apply/Lapply/Sapply/parallelize/parallelize_call to turn the parallelization
mechanism on or off. The global environment is modified by a call to this
function.}
\usage{
parallelize_setEnable(state)
}
\arguments{
\item{state}{TRUE or FALSE to enable or disable parallelization}
}
\value{
The returned value is undefined.
}
\description{
Turn on/off the parallelization mechanism

This function changes the definition of
Apply/Lapply/Sapply/parallelize/parallelize_call to turn the parallelization
mechanism on or off.

This function changes the definition of
Apply/Lapply/Sapply/parallelize/parallelize_call to turn the parallelization
mechanism on or off. The global environment is modified by a call to this
function.
}
\examples{

	parallelize_setEnable(FALSE);
	Lapply
	parallelize_setEnable(TRUE);
	Lapply

}
\seealso{
\code{\link{parallelize}}, \code{\link{parallelize_call}}, \code{\link{parallelize_initialize}}
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
