% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparallel.back.R
\docType{class}
\name{ParallelizeBackend-class}
\alias{ParallelizeBackend-class}
\alias{finalizeParallelization,ParallelizeBackend-method}
\alias{getResult,ParallelizeBackend-method}
\alias{initialize,ParallelizeBackend-method}
\alias{initScheduling,ParallelizeBackend-method}
\alias{isSynchroneous,ParallelizeBackend-method}
\alias{lapply_dispatchFinalize,ParallelizeBackend-method}
\alias{lapply_dispatch,ParallelizeBackend-method}
\alias{lapply_results,ParallelizeBackend-method}
\alias{parallelize_backend,ParallelizeBackend-method}
\alias{performParallelizationStep,ParallelizeBackend-method}
\alias{pollParallelization,ParallelizeBackend-method}
\alias{restoreParallelizationState,ParallelizeBackend-method}
\alias{saveParallelizationState,ParallelizeBackend-method}
\alias{scheduleNextParallelization,ParallelizeBackend-method}
\alias{finalizeParallelization}
\alias{getResult}
\alias{initialize}
\alias{initScheduling}
\alias{isSynchroneous}
\alias{lapply_dispatchFinalize}
\alias{lapply_dispatch}
\alias{lapply_results}
\alias{parallelize_backend}
\alias{performParallelizationStep}
\alias{pollParallelization}
\alias{restoreParallelizationState}
\alias{saveParallelizationState}
\alias{scheduleNextParallelization}
\title{Class \code{"ParallelizeBackend"}}
\description{
Base class for parallelization backends. Please refer to documentation of the methods
individually for more complete documentation.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ParallelizeBackend", config, signature)}. %% ~~ describe objects
here ~~ Config is a list containing parameters and signature is a character
string that uniquely identifies the computation that is to be parallelized.
}

\examples{

showClass("ParallelizeBackend")

}
\seealso{
%% ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ %% ~~or
\code{\linkS4class{CLASSNAME}} for links to other classes ~~~
\code{\linkS4class{ParallelizeBackendLocal}},
\code{\linkS4class{ParallelizeBackendSnow}},
\code{\linkS4class{ParallelizeBackendOGSremote}}
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\keyword{classes}
