% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparallel.back.local.R
\docType{class}
\name{ParallelizeBackendLocal}
\alias{ParallelizeBackendLocal}
\alias{ParallelizeBackendLocal-class}
\alias{initialize,ParallelizeBackendLocal-method}
\alias{lapply_dispatchFinalize,ParallelizeBackendLocal-method}
\title{Class \code{"ParallelizeBackendLocal"}}
\description{
Backend class implementing local execution.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("ParallelizeBackendLocal", config, ...)}.
During normal operation you do not have to create objects of this class yourself. Instead, \code{parallelize_initialize} will create such instances for you. The class can be configured with the following field in the \code{Lapply_config} argument of \code{parallelize_initialize}.
\itemize{
  \item freezerClass: defaults to \code{LapplyPersistentFreezer}
  \item stateDir: directory to store results from computations. This location is passed to \code{LapplyPersistentFreezer}. If temporary behavior is desired it can be set to: \code{sprintf('\%s/tmp/remote', tempdir())}.
   \item sourceFiles: a vector of files to be sourced prior to parallel execution
}
}

\examples{

showClass("ParallelizeBackendLocal")

}
\seealso{
\code{\linkS4class{ParallelizeBackend}},
  \code{\linkS4class{ParallelizeBackendSnow}},
  \code{\linkS4class{ParallelizeBackendOGSremote}}
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\keyword{classes}
