% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rparallel.back.R
\docType{class}
\name{ParallelizeBackendSnow-class}
\alias{ParallelizeBackendSnow-class}
\alias{initialize,ParallelizeBackendSnow-method}
\alias{lapply_dispatchFinalize,ParallelizeBackendSnow-method}
\title{Class \code{"ParallelizeBackendSnow"}}
\description{
Backend class for parallelization on SNOW clusters
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
	\code{new("ParallelizeBackendSnow", config, ...)}.
During normal operation you do not have to create objects of this class yourself. Instead, \code{parallelize_initialize} will create such instances for you. The class can be configured with the following field in the \code{Lapply_config} argument of \code{parallelize_initialize}.
\itemize{
  \item freezerClass: defaults to \code{LapplyPersistentFreezer}
  \item stateDir: directory to store results from computations. This location is passed to \code{LapplyPersistentFreezer}. If temporary behavior is desired it can be set to: \code{sprintf('\%s/tmp/remote', tempdir())}.
   \item sourceFiles: a vector of files to be sourced prior to parallel execution
   \item libraries: a vector of package names to be loaded prior to parallel execution
   \item localNodes: an integer number of how many parallel snow jobs are to be created. This should not be larger than the number of (logical) cores available as a general rule. A snow cluster is created using the \code{makePSOCKcluster}
}
You should be able to run a so-called \code{PSOCKS} cluster to use this package. See the \code{parallel} package for details (see also).
}
\examples{
showClass("ParallelizeBackendSnow")

Lapply_config = list(parallel_count = 24, backends = list(
    snow = list(
      localNodes = 8, sourceFiles = c('myScript.R'), libraries = c('boot')
    )
);
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
\code{\link{makePSOCKcluster}},
\code{\linkS4class{ParallelizeBackend}},
\code{\linkS4class{ParallelizeBackendLocal}},
\code{\linkS4class{ParallelizeBackendSnow}},
\code{\linkS4class{ParallelizeBackendOGSremote}}
}
\keyword{classes}

