% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparallel.R
\docType{class}
\name{LapplyExecutionState-class}
\alias{LapplyExecutionState-class}
\alias{LapplyExecutionStateClass}
\title{Class \code{"LapplyExecutionState"}}
\description{
An instance of this class reflects the entire lifetime of a dynamic
parallelization.
}

\section{Extends}{


All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}

\examples{

showClass("LapplyExecutionState")

}
\seealso{
\code{\link{LapplyFreezer-class}}, \code{\link{LapplyState-class}}
}
\author{
Stefan Böhringer

Maintainer: Stefan Böhringer <r-packages@s-boehringer.org>
}
\keyword{classes}
