\name{LapplyPersistentFreezer-class}
\Rdversion{1.1}
\docType{class}
\alias{LapplyPersistentFreezer-class}

\title{Class \code{"LapplyPersistentFreezer"}}
\description{
Subclass of \code{LapplyFreezer} that stores results on disk. See \code{LapplyFreezer} for more documentation.
}
\section{Extends}{
Class \code{"\linkS4class{LapplyFreezer}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
	\code{\link{LapplyFreezer-class}}
}
\examples{
showClass("LapplyPersistentFreezer")
}
\keyword{classes}
