% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rparallel.R
\docType{class}
\name{LapplyPersistentFreezer-class}
\alias{LapplyPersistentFreezer-class}
\alias{LapplyPersistentFreezerClass}
\title{Class \code{"LapplyPersistentFreezer"}}
\description{
Subclass of \code{LapplyFreezer} that stores results on disk. See
\code{LapplyFreezer} for more documentation.
}
\section{Extends}{
 Class \code{"\linkS4class{LapplyFreezer}"}, directly.

All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("LapplyPersistentFreezer")
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
\code{\link{LapplyFreezer-class}}
}
\keyword{classes}

