\name{LapplyProbeState-class}
\Rdversion{1.1}
\docType{class}
\alias{LapplyProbeState-class}

\title{Class \code{"LapplyProbeState"}}
\description{
This subclass of \code{"\linkS4class{LapplyState}"} tracks probing runs of the parallelization process.
}

\section{Extends}{
Class \code{"\linkS4class{LapplyState}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.
}


\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\note{
See documentation of \code{"\linkS4class{LapplyState}"}.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{"\linkS4class{LapplyState}"}
}
\examples{
showClass("LapplyProbeState")
}
\keyword{classes}
