\docType{class}
\name{LapplyProbeState-class}
\alias{LapplyProbeState-class}
\title{Class \code{"LapplyProbeState"}}
\usage{
  LapplyProbeStateClass(...)
}
\description{
  This subclass of \code{"\linkS4class{LapplyState}"}
  tracks probing runs of the parallelization process.
}
\note{
  See documentation of \code{"\linkS4class{LapplyState}"}.
}
\section{Extends}{
  Class \code{"\linkS4class{LapplyState}"}, directly.

  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("LapplyProbeState")
}
\author{
  Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
  \code{"\linkS4class{LapplyState}"}
}
\keyword{classes}

