\name{LapplyRunState-class}
\Rdversion{1.1}
\docType{class}
\alias{LapplyRunState-class}

\title{Class \code{"LapplyRunState"}}
\description{
This subclass of \code{"\linkS4class{LapplyState}"} tracks running of code of the parallelization process.
}
\section{Extends}{
Class \code{"\linkS4class{LapplyState}"}, directly.

All reference classes extend and inherit methods from \code{"\linkS4class{envRefClass}"}.

}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\linkS4class{LapplyState}}
}
\examples{
showClass("LapplyRunState")
}
\keyword{classes}
