\docType{class}
\name{LapplyRunState-class}
\alias{LapplyRunState-class}
\title{Class \code{"LapplyRunState"}}
\usage{
  LapplyRunStateClass(...)
}
\description{
  This subclass of \code{"\linkS4class{LapplyState}"}
  tracks running of code of the parallelization process.
}
\section{Extends}{
  Class \code{"\linkS4class{LapplyState}"}, directly.

  All reference classes extend and inherit methods from
  \code{"\linkS4class{envRefClass}"}.
}
\examples{
showClass("LapplyRunState")
}
\author{
  Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
  \code{\linkS4class{LapplyState}}
}
\keyword{classes}

