% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparallel.R
\docType{class}
\name{LapplyRunState-class}
\alias{LapplyRunState-class}
\alias{LapplyRunStateClass}
\title{Class \code{"LapplyRunState"}}
\description{
This subclass of \code{"\linkS4class{LapplyState}"} tracks running of code
of the parallelization process.
}

\section{Extends}{
 Class \code{"\linkS4class{LapplyState}"}, directly.

All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}

\examples{

showClass("LapplyRunState")

}
\seealso{
\code{\linkS4class{LapplyState}}
}
\author{
Stefan Böhringer <r-packages@s-boehringer.org>
}
\keyword{classes}
