\docType{class}
\name{ParallelizeBackendOGSremote-class}
\alias{getResult,ParallelizeBackendOGSremote-method}
\alias{initialize,ParallelizeBackendOGSremote-method}
\alias{initScheduling,ParallelizeBackendOGSremote-method}
\alias{lapply_dispatchFinalize,ParallelizeBackendOGSremote-method}
\alias{ParallelizeBackendOGSremote-class}
\alias{performParallelizationStep,ParallelizeBackendOGSremote-method}
\alias{pollParallelization,ParallelizeBackendOGSremote-method}
\title{Class \code{"ParallelizeBackendOGSremote"}}
\description{
  Backend class supporting Open Grid Scheduler support on
  remote machines
}
\section{Objects from the Class}{
  Objects can be created by calls of the form \code{}. %%
  ~~ describe objects here ~~
}
\examples{
showClass("ParallelizeBackendOGSremote")
}
\author{
  Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
  %% ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ %%
  ~~or \code{\linkS4class{CLASSNAME}} for links to other
  classes ~~~ \code{\linkS4class{ParallelizeBackend}},
  \code{\linkS4class{ParallelizeBackendLocal}},
  \code{\linkS4class{ParallelizeBackendSnow}},
  \code{\linkS4class{ParallelizeBackendOGSremote}}
}
\keyword{classes}

