\docType{class}
\name{ParallelizeBackendSnow-class}
\alias{initialize,ParallelizeBackendSnow-method}
\alias{lapply_dispatchFinalize,ParallelizeBackendSnow-method}
\alias{ParallelizeBackendSnow-class}
\title{Class \code{"ParallelizeBackendSnow"}}
\description{
  Backend class for parallelization on SNOW clusters
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("ParallelizeBackendSnow", config, ...)}. %% ~~
  describe objects here ~~
}
\examples{
showClass("ParallelizeBackendSnow")
}
\author{
  Stefan Böhringer <r-packages@s-boehringer.org>
}
\seealso{
  %% ~~objects to See Also as \code{\link{~~fun~~}}, ~~~ %%
  ~~or \code{\linkS4class{CLASSNAME}} for links to other
  classes ~~~ \code{\linkS4class{ParallelizeBackend}},
  \code{\linkS4class{ParallelizeBackendLocal}},
  \code{\linkS4class{ParallelizeBackendSnow}},
  \code{\linkS4class{ParallelizeBackendOGSremote}}
}
\keyword{classes}

